Clear neighbor tables in ensemble dtor
Summary:
SAI tests on Tajo can hit a condition where removing a subnet
route fails because a neighbor in that subnet is considered to depend
on the subnet route. To avoid this for now, remove all the neighbors
as well as the routes and intf addrs. SaiSwitch handles the neighbor
delta before the route delta, so this behaves as desired.

Differential Revision: D20528090

fbshipit-source-id: 3f224fb17aaa6412c2a5ddb97dd68f851291b581