recreate sai table manager in fake post warmboot
Summary:
In order to fake a warmboot, some tests releases and reloads all the SaiObjects in the SaiStore. This will cause new SaiObjects to be created in the SaiStore. The liveness of all SaiObjects that were created pre-warmboot will be set to False. When the manager table is destroyed, the Api call to the adapter on these SaiObjects will be no-op.

For instance, Sai Adapters has a dependency check which does not allow us to delete a scheduler before resetting the scheduler OID on a queue to NULL. To handle this, before destroying a scheduler object, I tried resetting the OID of a scheduler to NULL OID on a Queue object but this causes these warmboot tests to fail. The reason being the sai objects are not live.

This diff is to prevent calling the manager table destructor and leak memory very similar to the hw tests - https://fburl.com/diffusion/p48u6853

Reviewed By: phshaikh

Differential Revision: D20770589

fbshipit-source-id: 99b223e60f07fa334b4ca8008566b674018871b3