QCM firmware load changes.
Summary: High level Idea is to load  firmware in ucontroller when switch is getting initialized. Firmware is needed only on certain platforms and not all. Loading is gated with gflag. Changes mostly picked up from D19631940 with small modifications. These changes are done for QCM fw load only. Intent is that link scan fw changes can fit right in when they are ready.

Reviewed By: shri-khare

Differential Revision: D20761034

fbshipit-source-id: 19c999d8f52173cf884930c26a523a65f445739d