also log restart times to stderr
Summary:
This module helps us understand cross-process restart times a
bit more easily. We have been debugging a related issue in SPR pods,
but it looks like one of our counter refactoring efforts stopped these
showing up in ods anymore.

This adds the durations in to our logs as well, so it is slightly
easier to search back in time and figure out restart times without a
lot of scrolling and a little math.

Differential Revision: D20856685

fbshipit-source-id: 10d2eb227960464de2a3f4772732f52f4518b0b7