Clear watermark stat after reading.
Summary:
BRCM does not fully implement _ext versions of
get stats api on ports, queues. And hence we can't use READ_AND_CLEAR
to clear watermarks after reading them. Their recommendation
is for client application to explicitly clear the watermarks when then want.
So use that here.

This is not the best approach, as to get true watermarks you need a
atomic READ_AND_CLEAR (else the peak could be hit b/w the time of
read and clear, which will be missed by subsequent read), which is
what I think the h/w is capable of doing. I am still pushing brcm
to implement this on CS10178882.

Sending this out to start running our revised watermark tests from
D20819651. Will adapt the code as the CSP evolves

Differential Revision: D20854127

fbshipit-source-id: 3bc4f76aa2b7778b1708f3f3a886032c9733f415