avoid bcm_port_stat_enable_set on already enabled ports
Summary:
This api was found to leak resources internally. This adds a
workaround to avoid calling it on ports that are already enabled. This
assumes that enabled port implies that stats are there, but this
should be temporary while we get a better fix from broadcom.

D20894472 was a potential fix at the bcm layer.

Differential Revision: D20899745

fbshipit-source-id: a4120959e596f4c5645a3139d872563c4aeafa75