LookupClassUpdater: Maintain subnets of interest
Summary:
Queue-per-host for l2 entry and neighbor entry is already implemented.
This is towards supporting queue-per-host for route entries (e.g. ip-per-task).

Maintain a list of subnets of interest. See comments in the code for why.
Subsequent diffs in this stack will build on this and make use of this.

Differential Revision: D20830839

fbshipit-source-id: 642b810acb5a54f3945b6fa6ceee3493ff4d3df9