LookupclassUpdater: Maintain IP to classID mapping
Summary:
Queue-per-host for l2 entry and neighbor entry is already implemented.
This is towards supporting queue-per-host for route entries (e.g. ip-per-task).

To determine classID for newly added route, we need to find classID associated
with its nexthop. Maintain IP to classID mapping so that we can look up this
info efficiently. Subsequent diffs in the stack would add logic that does this
lookup, this diff adds logic to maintain IP to classID mapping.

Differential Revision: D20834980

fbshipit-source-id: b9e9f48ed391fb0354dd5f50d292ee4b48a7bd90