Add a stress test to exercise stress stat enable work flow
Summary:
Added in response to T56148664, where we discovered
a resource leak on reenabling port stats. This adds a point test
to exercise that code flow.
This is not the most generic way of catching such issues, ideally
we would have something like a warm boot stresser. Today we get
that some what via cont canaries. Maybe we can add a warm boot
stress canary for each ASIC to warmboot 500 times back to back.

Reviewed By: aeckert

Differential Revision: D20904474

fbshipit-source-id: 487502c8066c9ea58b83ddeb2a5b1b3be65add87