Set configured switch state on warm boots
Summary:
For warm boots, in wedge_agent at this point we would
apply the config. Which if nothing changed would be a noop
In HwTests we don't save and apply the config on warmboot
and since the warm boot expectation is that switch came
back in a identical state. Mark the switch state is as
CONFIGURED

Differential Revision: D20911464

fbshipit-source-id: 91391067031efe42e25af8120fa61d1ded93f64b