QosMapApi
Summary:
In order to add support for QosPolicy in SaiSwitch, the first step
is to add support for the api at all. Add the new api, a new fake, and some
simple unit tests. Further tested higher up the stack via integrations with
store/ and switch/

A short description of the QosMapApi itself:
the two attributes for a qosmap of interest are TYPE and MAP_TO_VALUE_LIST
MAP_TO_VALUE_LIST is a list of pairs of unions, where each pair represents
a mapping from key to value for the corresponding object. e.g., if key is
dscp and value is tc, it is a dscp->tc mapping. TYPE determines which
union members should be filled in. We will be using:
SAI_QOS_MAP_TYPE_DSCP_TO_TC, SAI_QOS_MAP_TYPE_TC_TO_QUEUE (and eventually
extensions for EXP). So each QosMap object is a list of mappings of the
particular type, so we will be createing two (and then tree for exp) QosMaps.

Reviewed By: phshaikh

Differential Revision: D20312031

fbshipit-source-id: ce3b4320879c245550e9e79650a8a8f2b868efa4