LookupClassUpdater: Maintain Nexthop to Routes map
Summary:
Queue-per-host for l2 entry and neighbor entry is already implemented.
This is towards supporting queue-per-host for route entries (e.g. ip-per-task).

Routes inherit classID of its nexthop. Thus, when a classID is
associated/disassociated with a neighbor entry, we must update all routes whose
neext hop is the neighbor IP. To perform this operation efficiently, maintain a
nexthop to routes map.

Subsequent diffs in the stack use this.

Differential Revision: D20834981

fbshipit-source-id: d54108b5500b83bdbbbecd61555badd18819d48a