extend LogThriftCall to support future* apis
Summary:
I extended the class for logging thrift calls to also support
things if the handler chooses to implement the future* version of the
interface. Will add versions for SemiFuture and Coro as well, but this
was for use in the fboss pinger service, which still uses traditional
futures heavily.

Differential Revision: D20962679

fbshipit-source-id: 01cb48d5fa96b7f28cae559951c1ae7619214154