switch/SaiQosMapManager
Summary:
This diff adds support for programming the default QosPolicy defined
by SwSwitch in SaiSwitch. For now, I see no need to support multiple QosPolicy
objects, as we are just interested in dscp->queue. In SAI, we must use
dscp->tc and tc->queue, but in Cisco, tc->queue will always be the identity map.

Differential Revision: D20312030

fbshipit-source-id: a8a348fad5b5c7efb56f70abfa0e823f06bbcb63