stop asking hardware for fec status on thrift calls
Summary:
This breaks the dependency on the hardware for port related
thrift calls, which are the most common calls we get in
production. Having the dependency can lead to issues blocking thrift
cpu workers when we have hardware problems. This field is pretty
incomplete anyways and we need to support returning which fec is used
since it is not as simple as on/off.

This is a first mitigation step. We will follow up by:

1. Moving logic for deciding if fec should be enabled or not back up
in to SwitchState.

2. Replacing with a more descriptive enum using the platform to decide
what the state should be based on a port profile.

Differential Revision: D21043769

fbshipit-source-id: 3413b1c77348d3a7ba9e4b5646c35468bf08e6bf