Allow for a default value of a attribute
Summary:
Add plumbing so that we could pass in a callable (for
now I used a functor, but we can extend to passing in lambdas).
to get the default value of particular attribute. Immediate
use case for this is to handle where certain (optional) attributes
may not be implemented by a particular platform. In such a
case a get on them may return a NOT_IMPLMENTED error code.
We now provide a hook to allow our SaiApi layer code to
return a default value for such attributes. Downstream code
can then look at this default value and take appropriate action.

Use this facility for providing default values for global
qos maps since not all platforms support these.

Reviewed By: boryas

Differential Revision: D21034082

fbshipit-source-id: 579e24cfa96cad0a9b7c9326571ae73a7e1fb7db