Add pre-emphasis overwirte for AOI module
Summary:
We observed that AOI optics with certain PartNumber have set their RxPreEmphasis value as 2dB where other optic modules have been setting the value to 0.

This difference seem to have caused signal weaker on the channel. Thus we implement a remediation here is qsfp_service to overwrite the value.

Reviewed By: tfangit

Differential Revision: D21028634

fbshipit-source-id: 3866ee604f3219c9b44fcff696462fe7b6c6450f