extend LogThriftCall to support printing args
Summary:
Sometimes it is useful to print some of the parameters of a
thrift request. This adds support in the LogThriftCall library to do
that.

Implementation will be able to print anything that is serializable
with DynamicConverter.h, which should be basic types and collections
of basic types. Added an example in to LogThriftCall.h

   void dummyThrift(
       std::unique_ptr<std::string> a, std::vector<int> b, int c) {
     auto log = LOG_THRIFT_CALL(DBG1, *a, b, c);
     ...

which will output something like

  "received thrift call from x ... params: *a="hello",b=[1,2,3],c=99,

I also added a few optimizations in this diff:
  - LogThriftCall should be move only to avoid misleading destruction logs
  - Short-circuit all code if the log level is not actually on

Differential Revision: D21029461

fbshipit-source-id: 898ef4ca3160c230f5b3a5381036ad66d65e39d6