update concurrent indices in port manager
Summary:
As titled, removing this from vlan manager and moving it to port manager.

vlan member depends on "BridgePort", having to maintain this in vlan manager requires vlan member to depend on both "SaiPort" and "BridgePort" which makes  subscription mechanism complicated.

Deletion of SaiPort will trigger removal of BridgePort which triggers vlan member removal. But SaiPortID is not present in the context of subscriber for BridgePort.

Reviewed By: boryas

Differential Revision: D21146290

fbshipit-source-id: b323c690a53587e27adc80631f5274f27e096a9d