NextHopsAndRouteManager: maintain subnets of interest
Summary:
Queue-per-host for l2 entry and neighbor entry is already implemented.
This is towards supporting queue-per-host for route entries (e.g. ip-per-task).

Maintain a list of subnets of interest. See comments in the code for why.
Subsequent diffs in this stack will build on this and make use of this.

Differential Revision: D21105384

fbshipit-source-id: 77a922b7b02d1daaa02160e05719e37b6aa8d905