Helper to manage routes, and corresp nexthops & classIDs
Summary:
Queue-per-host for l2 entry and neighbor entry is already implemented.
This is towards supporting queue-per-host for route entries (e.g. ip-per-task).

Previous diff in the stack introduces outline for inheriting route classID from
first *reachable* nexthop. This is a diff towards managing data structures
needed for that in a separate class. It is just an outline.

Differential Revision: D21104017

fbshipit-source-id: 037bcc2fc61bed30d4830df2a0099464b0e11cf6