Towards inheriting route classID from first *reachable* nexthop
Summary:
Queue-per-host for l2 entry and neighbor entry is already implemented.
This is towards supporting queue-per-host for route entries (e.g. ip-per-task).

The current queue-per-host implementation for route entry inherits classID of
the first nexthop. However, it is possible that the first nexthop is not
reachable but second nexthop is. In such a case, with the current
implementation, traffic hitting the route will use second nexthop but hit
queue0.

To fix this, we need to change the semantics from 'inherit from first nexthop'
to 'inherit from first *reachable*' nexthop.

This is not trivial: this means that when a selected nexthop turns unreachable,
we must search if the corresponding route has another nexthop that is reachable
and inherit classID from it.

This is first diff towards implementing that logic, and has several
placeholders. Subseqeuent diff would add logic and unit tests.

Differential Revision: D21099268

fbshipit-source-id: 18a4f45da8387cf4c7baf5896e58f53fc3831630