HwSwitchEnsemble::ensureSendPacket[Switched|OutOfPort]
Summary:
We can't rely on the SAI send packet api to be synchronous. However,
our dataplane tests depend on knowing the packet went out before checking
counters. Rather than make every dataplane test implement a retry port stat
checking mechanism, this diff adds two tools:

One is a set of "ensureSendPacket" APIs which send a packet, but also check
the port stats before and after to ensure that at least some port stat went up.
This roughly simulates a synchronous send, but is quite expensive, so it is
only useful for tests that send a small number of packets.

The other is a more general "waitForPortStats" mechanism that allows one to
provide a lambda to repeatedly check against the fresh port stats with some
sleeps in between checks. ensureSendPacket is implemented in terms of this
mechanism. This is targeted at tests like the UDP blackholing test which
send such a large number of packets that it is better to wait on the final
condition as a whole, not each packet.

Differential Revision: D20951196

fbshipit-source-id: 419cf2fbbce486cf4f319566c9302ee7fd31ca0d