Honor interfaceMode from PortProfileConfig
Summary:
Previous in Minipack and Yamp, we used override getTransmitterTech() to get the expected InterfaceMode or Medium type to program PortResource. This is convenient and necessary because different vendor's gearbox hijack the the old-time connection b/w chip and optics. And all ports in Minipack and Yamp are the same.

However, when Wedge400 comes, we need to support both uplinks and downlinks design, which usually differs from each other because we mainly use optics for uplinks and copper cables for downlinks. Therefore, we can just set one value in this getTransmitterTech() override. Besides, Wedge400 has more requests on this media type settings based on speed. So tl;dr:
- The difference to use SR vs KR is on the optics CDR. If there is no CDR like 10G/40G optics, we use SR. Otherwise we use KR, which is >= 100G
- For downlink ports with copper cable, we always use CR

We've already introduced interfaceMode in PortProfileConfig. So we'll set the interfaceMode based on the profile config and won't use live optics transmitter tech to program port in TH3.

**NOTE** : this change will flap the port since we're changing the interface mode from the chip side. However, Minipack and Yamp ports are basically still untouched from this change with the same mode. The only change here is for Wedge400, and we don't have any prod Wedge400. So we're safe to roll out this change without using UU to coldboot agent.

Reviewed By: birdsoup

Differential Revision: D21307436

fbshipit-source-id: 07fd9c0169267d383ecfb50d74329e7087a857dc