Start to use folly log instead of glog in the lib
Summary: I2C bus based wedge platforms would do a I2C transaction to detect the presence of a transceiver, in which case there will be a line of i2c error in the log. This will be misleading and too verbose. Thus switch to it a deeper log level. Also start using xlog instead of glog since it's suggested.

Reviewed By: rajank7

Differential Revision: D21262292

fbshipit-source-id: ffd0aa33669ae26ace1853294f005c5e13af2bac