LookupClassRouteUpdater: outline to handle port/neighbor/route updates
Summary:
Queue-per-host for l2 entry and neighbor entry is already implemented.
This is towards supporting queue-per-host for route entries (e.g. ip-per-task).

Previous diff introduced state observer for LookupClassRouteUpdater.
It needs to handle updates to port, neighbor and routes and assign appropriate
classID for a route. Add an outlien for handling that. The actual
implementation would follow in the subsequent stacked diff(s).

Differential Revision: D21375372

fbshipit-source-id: 448ca828bd4b984b078d18b0c2af49a49ddd1ca1