LookupNextHopsAndRouteManager: data structures for classID computation
Summary:
Queue-per-host for l2 entry and neighbor entry is already implemented.
This is towards supporting queue-per-host for route entries (e.g. ip-per-task).

In order to implement 'route inherits classid of one of its reachable next
hops', we need to maintain several data structures. This diff introduces that.
See comments in the diff for detailed explanation.

Differential Revision: D21106095

fbshipit-source-id: a5f80454d6b4ce4417434eef96d7db74e3e5201b