Fix segfault in bcm port stat update
Summary:
There is a race condition where stat updates and port enable/disable can occur at the same time, and a port counter will be freed by the state update before it is updated by the stats thread, leading to a segfault. This was made more common by adding the asic fec stats, since those take significantly longer to update since they're not sw-buffered counters, and so they widened the window for the race condition to occur.

This diff adds locking to the port stat updates so that different threads won't modify them at the same time. Right now it just uses the existing 'last stats' locking for the duration of any stats write. I think it might make the intent more clear if we wrap up all the stats fields into a single object and lock on that, but for now this is a quick fix to the problem.

Reviewed By: aeckert

Differential Revision: D21356962

fbshipit-source-id: 0565024b49a56cfe953b1b68bb678b6a0a5bf348