HwTest repro for S199340
Summary:
ALPM algorithm on TD2 in a nutshell
Pivots - Less specific prefixes stored in TCAM. For a packet destination
pivots are searched. Pivot in turn points to a bucket, containing upto
16 more specific prefixes and their nhops.
Buckets - container for more specific prefixes and their nhops. If a match
cannot be found in bucket prefixes, pivot nhop information is used.
Once a pivot's pointed to bucket is full, the following happens
- 2 pivots are created by left shifting the pivot mask bits. E.g. a pivot
  with prefix X:20::/64, might get split into X:2::/60 and X::/60. Goal of
  pivot splitting is to evenly balance the more specific prefixes amongst
  the 2 buckets.
- 2 buckets are created and new pivots are made to point to them.
- Pivot nhop information is obtained by looking up in a internal trie data
  structure for pivots and finding the best match information.

In S199340, we hit a bug where nhop information for pivot corresponding
::/0 was not recovered properly on warmboot. Which made the following
sequence of events problematic post warmboot
 - New prefixes were learnt causing a pivot split
 - The new pivot inherited its nhop information from the default pivot i.e.
the best match for the new pivot's prefix was the default v6 route.
 Bucket capacity on TD2 is 16 prefixes.

The following test tries to recreate the scenario by
 - Adding 15 routes before warmboot. So these all fit into a single bucket/pivot
 - Adding 5 more routes post warmboot, leading to a bucket split.
 - Sending a packet to destination that would fall under the new pivot
prefix. As long as the bug exists, these packets would be dropped.

Reviewed By: phshaikh

Differential Revision: D21410219

fbshipit-source-id: 88ea1681670614e89423b402ac6ed6cb603dfb5a