enable truncate for ingress samplers for tomahawk3
Summary:
ingress sampling on all ports with sflow require truncate. this is because collectors sit on devices connected with lesser MTUs. in absence of this, larger samples, samples larger than MTU of device on which collectors are running, do not get all the samples.

truncate is allowed only on tomahawk3, accordingly set truncate for configure_ingress_sflow_mirror_sampling patcher which is used by both dc-flow and dc-controller.

Note that dc-flow runs all layers except RSW and FSW and these layers have TH3 only

For dc-controller, only minipack pods are being used which are again TH3.

Differential Revision: D21492965

fbshipit-source-id: e8b3e6d023d063448211f964f0e4bd23802d6a71