attempt to resolve mirrors if mirrors are updated
Summary:
an attempt is made to resolve mirrors only if neighbor or routes are updated. do this even if mirrors are updated. this is to ensure that if non-resolution affecting attributes (such as dscp or truncate etc.) is updated mirror gets re-resolved quickly

this is to fix a bug or behavior, in which mirror config is updated such as dscp or truncate, which shouldn't affect resolution parameters of mirror but ends up un-resolving it for a long until either route or neighbor update triggers resolution.

Differential Revision: D21497963

fbshipit-source-id: 7c76e07d088d3616bb242a3e10f1d2e1b5d66b9a