Flag to disable route portion of queue-per-host fix
Summary:
Queue-per-host for l2 entry and neighbor entry is already implemented.
This is towards supporting queue-per-host for route entries (e.g. ip-per-task).

This patch provides a mechanism to disable route portion of queue-per-host fix,
see comment in the code as to why.

This works on cold boot as well as warmboot.

Differential Revision: D21493708

fbshipit-source-id: 9cb4c4e516d127f65c81206e7862b7a7530093f8