LookupClassRouteUpdater: process route updates
Summary:
Queue-per-host for l2 entry and neighbor entry is already implemented.
This is towards supporting queue-per-host for route entries (e.g. ip-per-task).

This diff processes route updates. For new route, inherit classID from one of
the nexthops (if any). For route changes, if the nextHopSet changes, that may
trigger classID change as well.

Differential Revision: D21491863

fbshipit-source-id: 0fb7c4665a65de21205ab8deeacbc8a2b9751dd6