LookupClassRouteUpdater: Maintain subnets of interest
Summary:
Queue-per-host for l2 entry and neighbor entry is already implemented.
This is towards supporting queue-per-host for route entries (e.g. ip-per-task).

Maintain a list of subnets of interest. See comments in the code for why.
Subsequent diffs in this stack will build on this and make use of this.

Differential Revision: D21441250

fbshipit-source-id: 2ac8892977d5ec6cf306c45dd54699a73d7d10fe