lock based chunking concurrency in SaiSwitch
Summary:
After extensive discussion, our conclusion for an approach to safe
and simple concurrency in SaiSwitch is the following:
* continue using a single global mutex (saiSwitchMutex_)
* don't extend the proliferation of ConcurrentIndices (in fact, we should aim
  to sunset these)
* use finer grained locking/unlocking of the global mutex in update thread
  state changed operations to allow the necessary responsiveness for rx,
  link scan, stat collection etc.

This diff is the first proposal for implementing that design. Rather than
holding the global lock for the duration of stateChangedLocked, we now capture
the lock by reference in each StateDelta forEach[Added|Removed|Changed] lambda
and lock/unlock around the individual delta item unit of work. That way, each
route, neighbor, vlan, port, etc.. is handled atomically, but linear
effects like a large dump of routes or 128 expensive port operations can't
block other threads unduly.

Note that we opted for threading the std::mutex down into the delta processing
rather than something like coroutine based cooperative multitasking because
we felt the OS scheduler was the best entity to decide fairness between our
threads. If that turns out to be false, we may want to revisit those more
explicit priority based strategies again in the future.

Differential Revision: D21310388

fbshipit-source-id: c654148528cc165c85e0c25938f74d5599cba7b0