Return fec mode from getPortStatus
Summary:
Return fec mode in thrift calls so we can print the correct value in `fboss port detail`.
As of now, we display the wrong value if we're using any of the other fec types since the test is just checking for ON.
I've kept both fecEnabled and fecMode in the response until fecMode is fully rolled out and we can remove fecEnabled if it is not needed anymore.

Differential Revision: D21513705

fbshipit-source-id: a7417ab3baaa1df454a316c95f4bbfb6d16a366a