ecmp shrink on link down
Summary:
When we get a link down callback, we cannot afford to wait until we
signal up to SwSwitch and it communicates state back down to SaiSwitch. For
this reason, we must shrink the ECMP group in place. To this end, we already
have a thread of dependency subscriptions going from fdb entry up to next hop
group member: nhgm -> nh -> nbr -> fdb. So it is sufficient to delete the fdb
entry and the chain of dependent objects will be deleted as well.

Reviewed By: phshaikh

Differential Revision: D21310390

fbshipit-source-id: 06eddd59b80db563ad3446240efee9b656eb0104