Rename BencmarkMain to HwBenchmarkMain and set num_iters= 2^1
Summary:
As titled. This names the benchmark main appropriately
since we do set some custom settings based on whether this
is benchmark to be run on switch HW or not. In particular
num_iters is set to 1, telling folly benchmark lib to only
do one iteration of the benchmark. For nearly all hw
benchmark this is already defacto the case, since max
time to spend on a benchmark is set to 1 sec
- https://our.intern.facebook.com/intern/diffusion/FBS/browse/master/fbcode/folly/Benchmark.cpp?commit=e9cfc35915c85635c1066b025d1239e793fdf8d0&lines=66
However for micro benchmarks like ECMP shrink, benchmakring
time is under 25ms, so the framework will try to run it
multiple times. To get this to work we have to ensure that
- At the start of the run hw is setup in a clean state, (e.g. if
you are measuring route adds, routes added in the previous iteration aree
blown away).
- Hw is not torn down and setup in the n+1 iteration - since most
of our vendor SDKs are not capable of being torn down and resetup
without a process exit.

This undo adds significant unnededd complexity to our code and
is also not inline with our usual workflows. So just set num_iters
to 2^1 so we iterate just once (2^0 to 2^1, folly::Benchmark counts
in powers of 2).

Reviewed By: shri-khare

Differential Revision: D21560480

fbshipit-source-id: 2d4434ace191ec17f2d6534b08ec4ff55665abab