Track run state in HwSwtichEnsemble
Summary:
SwSwitch does this for wedge_agent have ensemble track this
for tests in benchmark. An alternative would be to move this run state
entirely to HwSwitch, but there are likely other dependencies in SwSwitch
that may want this even before HwSwitch is setup. So leaving here for
now.
Motivation for this two fold
- Folly may run a benchmark multiple times and we don't really want to
call hwSwitch->switchRunStateChanged multiple times. The problem
particularly comes when calling this API with transition to CONFIGURED
a second time. This causes SaiSwitch to restart the bottom half threads
which blows up like so: P130872532. We can ofcourse make point fixes
to that and check for threads already being created, but this lines up
more with the way we do things in wedge_agent/HwTest
- Get rid of initComplete_ flag

Reviewed By: phshaikh

Differential Revision: D21554697

fbshipit-source-id: 1f232492b87b042a058ae303371ff14dff832f37