link scan thread in SaiSwitch
Summary:
now that fdb manager supports handling a port down event, hook it up
to SaiSwitch. In the same style as BcmSwitch (and SaiSwitch RX) push the work
onto a bottom half thread and immediately return from the callback. This avoids
a real SAI re-entrancy deadlock in leaba sdk (same story as BcmSwitch, I
believe). Use the global lock to make it safe to execute the critical call on
the FdbManager

Reviewed By: phshaikh

Differential Revision: D21310389

fbshipit-source-id: 1cdb5c91a2ecb6df3d25ae9f2ac2ef9a0c9d7a71