fix build
Summary:
I broke build of this in D21640468. I had to change type() to
getType() because type is a reserved keyword.

I did some fbgs and nobody seems to use this api and it is for
distinguishing sfp vs qsfp and we haven't used sfp in years.

Fixing forward by changing handler properly.

Reviewed By: shri-khare

Differential Revision: D21653715

fbshipit-source-id: 041e4fdc2a7c0dd70b87cf3ff5d443c7d3e20429