Do a full comparison on neighbor change, before deciding that this is a noop
Summary:
Previously we were just concluding that when both old and
new nbr entries are pending there is nothing to do. However this
is not quite true. Within a pending entry there could be
a field change (e.g. dst mac, rif etc) that needs to be
reflected in HW. So do a deeper comparison for such cases

Reviewed By: phshaikh

Differential Revision: D21734755

fbshipit-source-id: ebd7f543d8dd68716bb9bd46fc04fbb4007ea6cf