neighbor as subscriber of SAI port
Summary:
This allows neighbor to know the associated port status while neighbor is being created. In general a neighbor may exist even if link is down because state delta for port state change is not yet applied. In such situation, neighbor must see if port state is down and if so prevent notifying next hops and next hop group members.

This handles a situation
- link down
- resets fdb, neighbor, next hop, next hop group member (shrink)
- warm boot (before state delta could get applied)
- since sw switch neighbor entry still exists, and SAI port also exists, fdb, neighbor, next hop and next hop group will expand even if link is down

if neighbor subscribes to port, it can look at link state and notify next hop only if port status is up

Differential Revision: D21735919

fbshipit-source-id: d167cc5748002f1e87ec2002026199b487b0df87