Add polling of prbs stats into StatUpdater
Summary:
In order to have a longer BER monitoring period, we need to poll the error count and calculate BER with the port speed. So here we add it to the update thread in StatUpdater.
We only do this poll on the ports who has PRBS enabled. And the adding and removing from the mapping is done through refreshPrbsStats function.
For calculating BER, we need to get the port speed, number of lanes and lane rate through profile.

Reviewed By: tfangit

Differential Revision: D21378687

fbshipit-source-id: 57a0804ca47d2e1ac053972a86668255023324a2