add methods for cleaner creation of bcm test configs
Summary:
Right now, tests that modify configs usually just directly access the array of ports by index instead of by ID, and there's a lot of duplicated code for setting fields in the port config.

This just adds some helper methods to make this cleaner (and to support the later refactoring of configs to include all ports).

Differential Revision: D20169806

fbshipit-source-id: 8e33fc93314891be963eaace6b9e072520bf19c9