return aggregatePortId from fetchL2Table()
Summary:
Previously we return the hw trunk group id via thrift, but
this is not particularly helpful to the user, since the id is
unknown. This makes it difficult to evaluate if we are successfully l2
learning on an aggregate port.

This had been intentional to avoid a race given that we could access
the BcmTrunkTable from the thrift thread and the update
thread. However, the race is not really an issue in prod since we only
mutate the trunk table during the applyConfig step. We should still
think about properly synchronizing things, potentially thinking about
disallowing thrift calls to directly hit BcmSwitch apis from the
thrift worker threads.

Differential Revision: D21387151

fbshipit-source-id: 3760eeab312fae51bc0ff835d501a2f6c15eb705