make BcmTrunkTable id lookups thread safe
Summary:
We want to be able to do id lookups from other
threads. Specifically, when we populate the l2 table via thrift, we
want to return the configured AggregatePortId, not the underlying
bcm_trunk_t.

In order to support this we now manage a separate id lookup map inside
of BcmTrunkTable that is a folly::ConcurrentHashMap.

Differential Revision: D21819816

fbshipit-source-id: 242d6c56fcea010ccc23c630fb2043f2afaad07a