Towards porting more Bcm ecmp tests to be Hw agnostic
Summary:
As titled. Have more BcmEcmp tests use the Hw
agnostic abstractions. Mostly a simple port, but I got
rid of the check where we were looking at sw egressids
We are not bringing down links in runSimpleTest, so the
sw ECMP members are the same as HW ecmp members, so no
need to double check. If we ever feel the need we can
also build a getEcmpMembersInSw abstraction, but seems
like a overkill for now.
I also removed a couple of basic ECMP and UCMP tests.
Both these are already tested multiple times in
HwEcmpTests and HwLoadBalancerTests

Reviewed By: phshaikh

Differential Revision: D21776629

fbshipit-source-id: 6fed3310df901cb87830b8286f5a244a64f83be7