Kill Hw tests that should have just been UTs
Summary:
These tests are a mirror of what we (almost)
already test at in UTs
https://github.com/facebook/fboss/blob/master/fboss/agent/state/tests/RouteTests.cpp#L3104-L3142
I say almost, because in UTs we are testing with 4 nhops vs in
HW test we were testing with both 4 and 8 nhops (TODO: extend UTs
to do 8 nhops as well).

At the HwTest level we should just test that the ECMP gets
programmed according to the Hw weights passed down. The actual
weights that get passed down post reduction, should be tested
via UTs. We already have many tests that verify that Hw programs
ECMP groups according to the weights (tested in {Hw, Bcm}EcmpTest,
HwLoadBalancerTests)

Differential Revision: D21780698

fbshipit-source-id: 6ed3f6541a246a57ee909f0cf4529f5c62a3c582