Update sw state by calling linkStateChanged after linkscan is enabled
Summary: This is a fix for the problem discussed in T63997855. Theoretically this was supposed to be a drop in replacement for the block in D20571397. However removing that block makes some tests fail that I couldn't fix. It shouldn't hurt to have both checks so, in the interest of fixing the production issue I propose adding this until I can figure out how to safely remove the other block.

Differential Revision: D21650268

fbshipit-source-id: 311b42f37a1648a08b57a7c2ee1bf617e7b12064