FakeSai exclude cpu port from getObjKey/Count for port
Summary:
This is consistent with what we get for platforms
that return port keys in that only front panel port keys
are returned.
Why does it matter?
- Keeping FakeSai consistent with the vendor adapters is
important
- Not excluding CPU port causes us to run into problems like
the code here
https://github.com/facebook/fboss/blob/master/fboss/agent/hw/sai/switch/SaiPortManager.cpp#L501
We try to reconstruct cpu port by looking into the platform config,
but we don't model cpu port in platform config and hence this
blows up for Fake but works fine on actual HW.

Reviewed By: shri-khare

Differential Revision: D21921084

fbshipit-source-id: 4f887968c935708e521d5d310423e207afd5c712