Combine v4 and v6 in discard counter test
Summary:
As titled, the functionality tested here is small enough
to be able to combine w/o making the test onerous to read.
On the +ve side we save the cost of 2 extra hw test runs.
Further I am adding another variation of this for unresolved
nhops and doubling those tests for v4 and v6 seems wasteful

Differential Revision: D21924495

fbshipit-source-id: 71e7e211937d9e714c3bdb4f4ca3a390e489a50f