Program xphy interface type from ProfileSideConfig
Summary:
Currently we hard-coded the xphy system and line side interface type by:
```
  auto backplaneInterfaceMode =
      (side == Side::SYSTEM) ? bcm_pm_InterfaceCR : bcm_pm_InterfaceKR;
```
This actually doesn't match the soc from Acton. Besides, for SDK6.5.17, Broadcom recommended:

{F239209978}

So we should utilize the new PlatformMapping design to read port and external phy config from PlatformMapping and avoid hard coding.

Unfortunately, we used to think iphy and xphy system are identical. But it turned out, this is not always true. In iphy we use KR2 for lane interface mode, here we use AUI_C2C for system side interface. So I introduced `xphySystem` in PortProfileConfig (D21827220). Therefore, when we call ExternalPhy::programOnePort(), we should use this new `xphySystem` to program instead of `iphy`

Differential Revision: D21875576

fbshipit-source-id: 662f90e52e5d24e8536c928b98ed35de76bde4c9