First delta process UTs for QosMaps and bug fix
Summary:
As titled. Now that we have a SaiSwitch
start writing UTs that exercise SaiSwitch code.
First of these is for delta processing of Qos maps.
This in turn found a bug where by the sai switch
manager was not going through the expected add
interface on qos manager but directly creating
tc and dscp mapping and thus never creating a handle.
Fix that.

Reviewed By: phshaikh

Differential Revision: D21921559

fbshipit-source-id: 4ce17f2af59806957f2b313d009c0829628a358b