Handle pseudoWarmboot post test refactor in D21921264
Summary:
In D21921264 I refactored sai switch layer UTs
to create SaiSwitch and not just the manger table. However
I left the fake WB tests as a TODO. Finishing that up now.
From what I can see the fake WB tests really just test
the SaiStore WB sequence. So the ask the SaiStore to
do a WB exit and then reload it. This takes care of the
store. However now the managers are holding pointers to
released objects so we don't want to access (or delete
them anymore). Previously this was handled by leaking
the managerTable. Now since we actually create a SaiPlatform
as the root object. Leak that for the same effect

Reviewed By: phshaikh

Differential Revision: D21993518

fbshipit-source-id: dd2022a20e1be4b2eb21e9f992c883fe7dc11044