Pull up logic to create and destroy global qos maps
Summary:
As titled. This logic is common b/w port and switch
mangers and we were duping it (poorly). And infact there
was a bug where we were not removing the default qos policy
in PortManger when we were supposed to.

TODO: Add a way to UT these changes. Since UTs run against
the FakeAsic, which has all features turned on the QosMapManager
tests always follow the path for setting qos policy at the
Switch level

Differential Revision: D22012371

fbshipit-source-id: 6e16f557874303cdb2cf401cdbeedbbbc1441086